channel-selection-task: Add handling for Channel Preference Report.
As part of PPM-1973, we need to move the handling of the Channel
Preference flow to the appropriate tasks, in the agent's side this is
the Channel Selection Task.

When receiving the CHANNEL_PREFERENCE_QUERY_MESSAGE need to store its
information locally on the task and propagate a CHANNELS_LIST_REQUEST
VS-CMDU to each of the agent's radios.

Once the task has received the CHANNELS_LIST_RESPONSE from both its
radios, need to create and send a CHANNEL_PREFERENCE_REPORT_MESSAGE
back to the controller.

This logic is moved from the son_slave_thread, where it is currently
handled. The next commit will remove the redundant code.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>