tests: ChannelSelection: configure an SSID before checking channels
So far the test did not configure any SSID.

This means that the configured agents are supposed to teardown their
radios.
When this happens, we cannot get the radios channels with iw.

The test has been accidentally passing on the NEC device so far,
because the default VAPs are enabled.

Configure an SSID as part of the test, so that we can be sure the
radios are configured and enabled.

Fixes PPM-2039

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>