tools: device: configuration: enable the console before configuring
As can be seen in [1], if the console is not enabled yet before we try
to configure the device, the configuration script will fail (because
it won't get the prompt only by sending an interrupt).

In other cases sending a newline to enable the console is done by the
Serial class, before it expects the prompt.
In this case however we use 'expect_prompt_on_connect=False' because
of the leftover ubus processes (see PPM-2029), so no newline is sent.

As a workaround, first send a newline to enable the console if
needed. Wait for one second before sending the interrupt, to make sure
the interrupt does not stop login scripts from running (for example,
sending an interrupt too soon can prevent the prplOS prompt from being
set.

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/2346414128

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>