beerocks defines: add Channel Preference Constants
This is a preparative commit

To reduce the number of "magic numbers" in the code, need to use
constant values.
The Channel Preference Constants will hold the following constant
values: Invalid = -1, Non-Operable = 0, Lowest = 1, Best = 15.

These values will be used in the Dynamic Channel Selection task. and
the Preference handling functions in the DB.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>