DB: Add Preference Timestamp
This is a preparative commit.

Currently we do not have a system that sends unsolicited preference
reports, as mentioned in the Multi-AP specification:
> If a Multi-AP Agent’s channel preferences change, it shall send an
unsolicited Channel-Preference-Report to the Multi-AP Controller
indicating the Multi-AP Agent’s current preferences.

So to support the agent's preference updates, we can set an expiration
limit in the controller, and before checking the preference report we
need to check if it has expired or not.

- Add a timestamp to the latest update of the preference.
- If a preference report has expired, Send a new Preference-Query.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>