controller: don't ignore DHCP_COMPLETE_NOTIFICATION
dhcp_complete_notification is sent to the controller
without a radio mac. This is the expected behavior.
Currently, the message is ignored which leads to the
IP of the station not showing in the bml-connection-map.

Don't ignore this message even though radio_mac is zero.

PPM-2061

Signed-off-by: anantkumar <cchavda@maxlinear.com>
(cherry picked from commit 57bd6805d00af7101918e55f8c6e473d9f35f97f)
Signed-off-by: Frederik Van Bogaert <frederik.vanbogaert@mind.be>