pre association steering task: add multiple AP Configs support
Up until now, When adding a group, The requirements were to provide
both AP configuration of 2.4ghz radio, and AP configuration of 5ghz radio.
Now, This task will support a variable number of AP Configurations.
In practice, the number of AP Configuration can vary between one to three,
and each one of them can be either 2.4ghz or 5ghz.

Signed-off-by: Yoni Abecassis <yabecassis@maxlinear.com>