ci: certification: add a script to generate Gitlab CI yaml
With the number of EasyMesh revisions and the number of tests
increasing, adding more tests to the platform pipeline is cumbersome,
and results in a long loading time for the pipelines web page.

In order to make it easier to generate the yaml configuration for
certification tests, a python script is added.
For each revision, a template, hidden job should be created
beforehand (like it's already the case for R1 and R2). Then, the
script can be used to generate one job per test. To reduce the web
page loading time, the tests are grouped by base names.

In a follow-up commit, this script will be used to generate all the
individual certification tests.

Note that for ease of use, the job template has been sightly
changed (e.g. always mention the revision, change order of fields in
job names, ...).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>