ci: configuration: switch to using TR-181 network configuration
Starting with prplOS 1.4, TR-181 components have been added to manage
the network configuration (tr181-bridging, ethernet-manager,
ip-manager, etc).

The TR-181 bridge manager currently overrides some of the UCI bridge
settings. As a consequence, after a reboot the eth0_4 interface of the
axepoint-based devices is added back to the bridge.

To avoid this issue and prevent others in the future, do the network
configuration using the TR-181 datamodel, over ubus.

Notes:

- UCI sections (network interfaces) currently need to be created
beforehand to be able to assign them an IP address.

- Interface names are not yet propagated to UCI sections. This is
available in a later ip-manager version, but not the one we currently
have in feed-prpl (v1.11.1).

- VLAN interfaces are not supported yet in the ip-manager version we
currently have in feed-prpl (v1.11.1).

- Previously some certification devices didn't have a data IP yet (in
theory it's not required), one has been added.

Closes PPM-2068

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>