DCSr2: Send Preference Query
To support the Channel Selection flow, as part of PPM-1973, expand the
DCSr2 logic to create & handle the Channel Preference Query and Report.

If the existing Preference has expired we need to send a new Preference
Query to the appropriate agent.
The returning Channel Preference Report will also be handled in the
DCSr2 Task.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>