ci: certification: add and use template file for R1 tests
Add a template file containing the hidden jobs that are re-used by
each individual jobs. Include the files containing the individual
jobs.

These hidden jobs were previously located in the same file that
contained the individual test jobs, but placing it in a separate file
makes it easier to re-generate the job files.

Use the new template file by including it in 'ci/platform.yml'.

Since we changed the pattern of the job names, also adjust the script
section in '.certification-generic'.

Note that we don't do the same for the R2 tests, because the
individual R2 jobs are going to be removed (in favor of the R4 tests).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>