ci: certification: remove individual R2 tests
With the R4 testbed, we will be able to run R2 tests. To reduce the
overall number of CI jobs that we have and because we will most likely
disconnect the R2 testbed at some point anyway, the individual R2
tests are removed.

Note that the R4 testbed will also be able to run R1 tests. However,
the individual R1 tests are kept for now for two reasons:

- The R4 testbed doesn't have the ability to cover all the R1 tests
yet (e.g. the tests involving stations are not working yet).

- We have two R1 testbeds, which means they're good candidates to run
two certification tests as part of the merge train.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>