ci: certification: add back merge train tests
We now have a script to generate the individual test jobs for
certification tests.  We had two special tests (the tests that we run
as part of the merge train) that are not covered by the script, so add
them back manually.

Since they already have a "manual" equivalent covered by the script,
change the "rules" so that the test is only created within the merge
train.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>