tests: certification: add R4 tests
As the R4 test plan v4.0 does not match the tests that were defined in
the R4 UCC (e.g. 4.6.1 is "removed" in the test plan, but present in
the UCC - 4.6.2 is present in the test plan, but not in the UCC), this
list comes from "Test_Data_EasyMesh_R4_v1.0.xlsx", which was provided
by the WFA support team separately.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>