AgentDB: Realign the ChannelPreference structure
As part of PPM-1973

we need to realign the preference storage mechanism
in the Channel-Selection flow. Up to now we would store only the
controller's preference while using the channel-list as a reflection of
the radio's preference.

This is incorrect as we only need the controller's preference when
deciding on a Channel-Switch.

To better reflect this we need to store the radio's preference in the
agentDB, and store the controller's preference only while handling the
Channel-Selection-Request they are sent with.

- Remove the controller preference element from the AgentDB.
- Move the sChannelPreference structure to inside the sRadio structure.
- Add a channel preference map for the radio element.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>