Channel-Selection-Task: Handle the Channel-Selection feature
As part of PPm-1973, we want to move all of the agent's general
Channel-Selection logic from the son-slave-thread to the appropriate
task.

This is a preparative commit.

- Add get_preference_for_channel which returns the preference value for
a given channel in a given preference.
- Add get_cumulative_preference which returns the cumulative preference
for a given channel in a given preference.
- Add default values for sSelectedChannel
- Remove unused sPendingRadioPreference
- Add sIncomingChannelSelectionRequest to store the Channel-Selection
requests and the controller preference.
- Remove the m_expected_channel_selection as the Channel-Selection
feature should be handled in the Task and not forwarded to the agent.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>