Channel-Selection-Task: Send the Channel-Selection-Response
As part of PPm-1973, we want to move all of the agent's general
Channel-Selection logic from the son-slave-thread to the appropriate
task.

After parsing the Channel-Selection-Request we need response.
For each of the agent's radios we need to build a
Channel-Selection-Response TLV with the appropriate response code.
If a radio did not receive a Channel-Selection-Request, we need to
respond with a ACCEPT code, since no changes are made.

After sending the Channel-Selection-Response need to send a
Channel-Switch-Request according to the previously received preference.
Normally, when a channel switch is required, a CSA notification will be
received with the new channel setting which is when the agent will send
the operating channel report.
In case of only a tx power limit change, there will still be a CSA
notification which will hold the new power limit and also trigger
sending the operating channel report.
If neither channel switch nor power limit change is required, we need
to explicitly send the report.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>