boardfarm: Fix test for nbapi radio
With the recent changes made as part of PPM-1973, we saw that the NBAPI
radio test is failing because of misaligned expectations.

The test expected only a single Operating-Channel-Report TLV, but
according to the specification we need to send one Operating-Channel
Report TLV per Radio.

Before this change, the Agent did not send the TLV correctly (it sent
only one per Agent regardless of the radio quantity) so I had to change
this code to allow the new changes to pass the NBAPI-Radio test.

Secondly, there is no need to send preference records for channels that
are not part of the radio the preference belongs to.
In the 2.4Ghz preference no need for operating classes 115-130.
And in the 5Ghz preference no need for operating classes 81-84.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>