bpl: cfg_uci: segmentation fault reading wireless configuration
bpl_cfg_uci xpects that the "wireless.radioX.path" contains the substring "soc/".
It does not verify if the substring is found or not.
Some routers (as TPLINK Archer C7 V5) do not have that substring.

Verify if pos is equal to std::string::npos and using the correct
path present in o->v.string from the system.

PPM-2070

Signed-off-by: Dmytro Puz <d.puz@inango-systems.com>
Signed-off-by: Paulo Campos <paulo.campos@beegol.com>