db: Find central channel in preference
This is a preparative commit.

As part of the work done in PPM-1973 we needed to get the preference
value of a specific channel.
But in case of channels that belong to operating classes 128,129&130
We need to consider the central channel.

If the incoming channel is a beacon channel need to check the preference
using the central channel.
- Add a flag to distinguish between a central channel and a beacon
channel.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>