DCSr2: Need to consider central channels for selection.
To be aligned with the Channel Selection requirements (PPM-1973), we
need to create an On-Demand Channel-Selection event and send it to the
DCSr2 task for handling.

If we receive an on-demand request for a channel that belongs to a
band that uses central channels instead of beacon channels we need to
send the selection request using said central channels.

- Consider central channels when building the channel-selection request.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>