DCSr2: Realign to support Auto Channel Selection flow.
To support the Auto-Channel-Selection flow as part of PPM-1973, the
following changes are needed.

- Add the WAIT_FOR_SCAN state to include the Scan into the Selection
flow, during this state we wait for the Channel-Scan to finish.
Once it is finished we need to request a new Preference-Report.
- Add the SELECTION_ABORTED state to handle the timeout failures in
the Selection flow.
- Add handling of timeout in the Selection flow.
- Create sOnDemandAutoChannelSelectionRequest structure to support the
On-Demand-Auto request.
When we use this request the task will know how to handle the
preference report the controller sends as part of the Channel-Selection
requirements.
- Update the CHANNEL_SCAN_REPORT_WAIT_TIME_SEC define to use a
constexpr instead.
- Fix typos in the DCSr2 task.
- Remove agents that reached a timeout in the Scan flow.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>