son_management: update selection request handling allow auto.
To support the Auto-Channel-Selection feature as part of PPM-1973 we
need to expand the handling of the TRIGGER_CHANNEL_SELECTION_REQUEST
CMDU to support using channel zero .

When an On-Demand-Selection-Request is received with the channel set
to zero it means the request is intended as an Auto-Channel-Selection.
Need to sent it to the DCSr2 Task for handling.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>