ci: configuration: disable the TR-181 DHCP clients
As described in PPM-2094, the DHCP clients are enabled by default on
the devices, and on the Turris Omnia (which uses the WAN port, on
which the DHCP clients run by default) they caused a race condition
with the IP that we try to set manually.

Disable the new TR-181 DHCP clients on the devices like we used to do
in UCI previously. Using ubus to set "Enable" to "false" on the DHCP
clients doesn't work (the clients still assign IPs), so stop and
disable them completely instead.

On the Turris Omnia (which is currently the only device where we use
the WAN port for control), we also need to always set the IP instead
of just adding it if there is none. This is because in case the DHCP
client had the time to already assign an IP, we need to change its
value.

Fixes PPM-2094

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>