ap manager: set preference to minimal on an empty ACS_REPORT
When we receive the ACS_REPORT in the AP-manager we use the returning
values as the basis to form the preference score for the supported
channels.
But if for some reason we receive and empty ACS_REPORT we need to set
the channels' scores to the minimal value that is still Operable.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>