framework: tlvf: add tlvApRadioVbssCapabilities
This commit adds support for tlvApRadioVbssCapabilities, the first TLV
we need to support Virtual BSSs.

Note that the VBSS TLVs all have the same TLV type, but unique TLV
subtypes.

As this is the first VBSS TLV we are adding:

- A new enum for the subtypes, eVirtualBssSubtype, is added. It is of
type enum_class so that we can use eVirtualBssSubtypeValidate::check()
to check message subtypes when parsing TLVs (exactly like we do for
TLV types). In this commit, only one TLV subtype is
supported (AP_RADIO_VBSS_CAPABILITIES).

- parseNextTlv() is also adapted to parse the TLV subtype (in addition
to parsing the type like we would do for TLVs that do not have a
subtype) for us to know what class to create.

Note that this TLV and the next ones that will follow are based on the
latest draft specifications (Draft_EasyMeshVBSSSpecSubmission_prpl_2022-05-10
Unencumbered by WFA Confidentiality) and are still subject to change.

The current version of the specifications did not contain descriptions
for the new TLV type, so one was arbitrarily chosen.

Also update auto-generated files.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>