btl: Fix broker size
Currently the broker_buffer_size, which is used to allocate the
serialized message's buffer, does not take into consideration the
Metadata header that is used as part of the message.
This causes an overflow in the case of very large messages.

To fix this issue, need to expand the broker_buffer_size to take the
Metadata size into the size calculation.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>