framework: tlvf: CmduMessage: add support for TLV subtypes
The VBSS specifications use a single TLV type (0xDE in the current
draft specifications), and unique TLV subtypes to be able to
distinguish TLVs.

Add a new type of header, sSubtypedTlvHeader, that is used to store
the subtype in addition to the type and length.

Since the subtype is 2 bytes long, we need to swap it like we do for
the length.

In follow-up commits, the new struct will be used to parse new TLVs
that we know contain subtypes (based on the TLV type, we can know if a
TLV contains a subtype or not).

While at it, remove unneeded typedef for sTlvHeader.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>