framework: tlvf: add test for creating VBSS Capabilities Reponse
The test verifies that we can create a new TX CMDU, like we would do
when we want to send a new VBSS Capabilities Reponse.

After finalizing the TX CMDU, verify that we can create a new RX CMDU
based on the same buffer. Among other things, this part checks that
parsing the TLV subtype is done correctly (especially that swapping
works for example on x86), and that the correct TLV can be
obtained (ApRadioVbssCapabilities).

Then, the content of the RX CMDU is checked to make sure it matches
the TX one.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>