Docker: Update base openwrt hash
According to what written in here:
https://patches.linaro.org/project/oe-core/patch/1399289179-5662-1-git-send-email-koen.kooi@linaro.org/
It looks there is a problem with ext2_fs.h that happens when the header is present
at configure time, but not at compile time.
It could be because e2fsprogs or one of its dependencies changed so it gets rebuilt,
at the same time that libarchive gets built.
Since it's been some time since we last updated it anyway,  update the base prplOS
version that we build.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>