tools: docker: builder: include prpl profile when building toolchain
Latest prplOS fails to build due to the following libarchive compilation
issue on one of the prplMesh builder machines:

  libarchive/archive_entry.c:60:10: fatal error: ext2fs/ext2_fs.h: No such file or directory

It seems like libarchive looks for ext2fs headers during configure step,
and if it finds them it will expect to find them during compile step, or
on the rare occasion when they aren't it will fail.

This offending libarchive package/component dependency was introduced
into prplOS recently with LCM stack integration, where cthulhu
(container lifecycle manager) uses that library.

This libarchive build race issue was fixed in prplOS in feed-prpl, where
fixed version of libarchive package is provided. So in order to use that
fixed package version we need to use prpl profile when building the
toolchain as well.

Fixes: PPM-2122 PCF-638
Signed-off-by: Petr Štetiar <ynezz@true.cz>
Signed-off-by: Adam Dov <adov@maxlinear.com>