yaml: Remove Redundent information from VS-TLV.
As part of PPM-1973 we want to expand the sent Channel Selection
Request message to include Vendor-Specific information, this is done to
support the On-Demand feature requested by MaxLinear.

Because of how the Channel-Selection-Task handles the Channel-Selection
request there is no need to sent the channel, bandwidth & operating
class of the on-demand request.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>