docker: Add BSSIDs interfaces to the Docker file
Preparative commit.
The operational BWL types report the interface name as part
of AP-ENABLE event or any event that contains information about the APs
BSSIDs.
Until now we didn't use the interface name, but we will need it to
prevent traffic separation configuration until all BSSIDs interfaces are
validated to be in the bridge.

Add to the Dummy Docker file a setup of dummy BSSIDs interfaces, so in
the future, prplmesh will be able to query the OS if the interface
is in the bridge or not.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>