bwl-dummy: Genarate dummy AP-Enabled event
When working with real Hostapd, it generate AP-Enabled events which are
propagated to the Agent.
The Dummy BWL is missing that functionality.
This functionality is needed to later check that every enabled AP
interface is in the bridge.

Add generation of AP Enabled event when calling to the dummy
update_vap_credentials() function.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>