agent: Add an "iface_name" variable to the sBssid struct in the AgentDB
Preparative commit.
Add an "iface_name" variable to the sBssid struct in the AgentDB, so
when the Agent gets an event that may update the BSS information in the
AgentDB, the interface name will be stored.
This change will help later to iterate over all BSSIDs of radio, and
check that the BSSID interface is in the bridge.
Also, being able to print the interface name is a big plus.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>