tlvf: Add WIFI_CREDENTIALS_UPDATE_RESPONSE message
Add a response message for the APMANAGER_WIFI_CREDENTIALS_UPDATE_REQUEST
message which will return when the AP Manager applies the new interfaces
configurations.
In the message, add the number of configurable BSSIDs that are
configurable according to the Hostapd configuration file.

This is message is required for initializing a timer in the
autoconfiguration task which will take place in a future commit.

Added the Auto-Generated code too.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>