Agent: Add a file descriptor to an interface name conversion function
Add a function to the RadioManager in the Agent that converts a file
descriptor of a Fronthaul entity (AP Manager or Monitor) to the radio
interface name.

This function will later be used in the AutoConfiguration task.

Make the radio manager public so it could be accessible from the tasks.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>