Agent: Add new state and variables to the sConfigurationParams struct
Add WAIT_AP_CONFIGURATION_COMPLETE state to the AutoConfiguration task.
This state will wait until the BSSIDs configuration is complete and
configure the traffic separation only after all APs BSSIDs will be
validated to be in the bridge.

Note that the state SEND_MONITOR_SON_CONFIG was removed since it is not
used.

In addition, add some variables to the AutoConfiguration tasks
sConfigurationParams struct that are needed for that state.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>