Agent: Move message handlers to the AutoConfiguration task
Move the following message handlers from the son_slave file to the
AutoConfiguration task:
cACTION_APMANAGER_HOSTAP_AP_ENABLED_NOTIFICATION
cACTION_CONTROL_HOSTAP_VAPS_LIST_UPDATE_NOTIFICATION

Note that this commit breaks the code since the code moved as it without
any adaptations. The reason is that it would be much easier to see the
changes done to the handler in a different commit.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>