Agent: Adapt copied message handlers
Adapt the copied message handler from the previous commit to the
AutoConfiguration task where they are now located.

Note that on the HOSTAP_AP_ENABLED_NOTIFICATION, the BSS is being added
to the enabled_bssids list.

On the HOSTAP_VAPS_LIST_UPDATE_NOTIFICATION, the flag
received_vaps_list_update is being set at the end of the function.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>