Agent: Implement WAIT_AP_CONFIGURATION_COMPLETE state
Implement the AutoConfiguration task state
WAIT_AP_CONFIGURATION_COMPLETE.

The state waits until getting the AP-Enabled event on all BSSIDs.
Note that it is possible to not get the events if the AP is not
configured for example if the configuration is not changed from what
is configured now.
In that case, after the timeout, the task will send a BSS Update request
to retrieve all the BSSIDs information.
Afterward, it will check that the all radio BSS interfaces are inside
the bridge as a prerequisite for configuring the Traffic Separation
configuration on the device.

All the required initialization for the task state are were added at the
M2 message handler function.

In addition, add some clang-format changes that it is not clear how
they weren't in the code until now.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>