common: add vht central frequency function
As part of the channel selection we need to know the central frequency
for a given channel & bandwidth pair.

With the helper function get_vht_central_frequency we can easily get
the central frequency for a given channel & bandwidth pair regardless
of their band (5GHz or 2.4Ghz).

Later we would need to expand this function to also include 6Ghz.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>