channel selection: set center frequency for 2.4Ghz
Before sending the HOSTAP_CHANNEL_SWITCH_ACS_START message, we need to
fill the vht center frequency for a channel switch on bandwidth 40Mhz
and higher.
We can use the newly added get_vht_central_frequency function in the
wireless utils. It uses the channels_table_5g & channels_table_24g
tables to find and calculate the central frequency for a given channel
and bandwidth pair.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>