controller: reorder the checking of task's response timeout
Avoid calling the response timeout handler when pending MACs is empty.
This fixes a scenario when a response message has been received,
and as a result of that the the pendings_macs has been emptied,
and afterward execute() method has been called, but the response timeout
is reached, resulting in calling the response timeout handler, even
though we have already handled the response.

Signed-off-by: Yoni Abecassis <yabecassis@maxlinear.com>