agent: Initialize "msghdr" struct explicitly.
prplMesh compilation fails on a struct initialization in UGW9.x.
The struct is defined on a socket.h on an external library,
and it is written a little bit differently on the version of that
file on the tree, compared to what we have on Linux.

Change the initialization of the struct to explicitly initialize the
struct members.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>