controller: add identification to client disconnect msg
A client disconnect request message can be issued by multiple components,
such as beerocks_cli, pre_association_steering task, and more.
Upon receiving a client disconnection response message, the handling
can be different between the components that sent the request.
Thus, A source component identifier is added to the message.
In the next commit, the corresponding code will be added to the agent.

Signed-off-by: Yoni Abecassis <yabecassis@maxlinear.com>