framework: tlvf: add test for creating VBSS Configuration Report
The test verifies that we can create a new TX CMDU, like we would do
when we want to send a new VBSS Configuration Report.
Then, check that we can create an RX message based on the TX buffer
and parse it. After parsing it, verify that every field matches the
expected value.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>