ap manager: handle empty ranking table
As decribed in PPM-2187 we see that in NEC WX3000HP we are receiving an
empty ACS report.
This causes the Radio Preference Ranking to return with all channels
set to Non-Operable (0).
This is most likely due an old DWPALD version that is running in that
setup.

In order to ignore this issue, and allow pipeline tests to pass, we
need to manually set the multiap preference score to 1 (Lowest score)
inorder for the Channel-Selection tests to pass.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>