controller: finish optimal path when client steering is done
On the SEND_STEER_ACTION state, we start a client steering task,
and wait for the client steering task until it ends.
However, when the client steering task ends, we need
to finish the optimal path task. otherwise, the
optimal path task will go to WAIT_FOR_HANDOVER state,
which we want to avoid, as stated in its state comment.

Signed-off-by: Yoni Abecassis <yabecassis@maxlinear.com>