backhaul manager: introduce new mechanism for detecting disconnection
Backhaul connection link does not assure actual connection is alive up
to the controller.

New mechanism is introduced with utilizing latest message timestamp
from the controller.

If latest message is timeout, agent sends one heartbeat message.
HLE TLV is used to detect connection health.

This TLV needs to be replied from Controller in one second according to
standard.
Even if it is not designed to be used as heartbeat,
it can be utilized as this purpose. Because, there are not much other
messages which could be triggered directly from Agent.

In normal flow, Controller sends periodical messages to Agent
like discovery, ap link metrics query.

With each message from controller, BH manager updates the timestamp.
If it exceeds message timeout, heartbeat is send.
We are expecting a reply within one second.
If it exceed connection timeout, it is considered as disconnection.

Message timeout is defined as 70 Seconds.
Connection timeout is defined as 80 Seconds.

If we lower message timeout, more heartbeat messages will be send and
it will increase the traffic. But, it could be more reactive.

https://jira.prplfoundation.org/browse/PPM-2089
--
Signed-off-by: Cihangir Odabas <codabas@assia-inc.com>