ci: certification: add back merge train tests as manual jobs
In commit e35eca01dd9bb1757702e9192b39b10d59412236, merge train tests
were added manually in 'ci/certification/templates.yml'.

The intention was that the test with the same name in
'ci/certification/R1/' would allow to have a manually-triggered test
outside of merge train, while still running it automatically when it's
part of merge trains.

However, as the job name is the same in both places, the one in
'ci/certification/templates.yml' takes precedence, which means we
don't have the possibility to manually trigger those two tests outside
of merge trains.

To fix it, change the name of the merge train tests so that it doesn't
override the manual one anymore.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>