controller: db: Add task identifier for vbss_task
Currently there is no efficient way to fetch the task_id for a created
vbss task.

This is necessary to push an event to the vbss_task. Specifically a
MOVE event.

Based on existing work done with the bml_task, dhcp_task,
agent_monitoring_task, etc, a vbss_task_id was declared in the db object
along with the relevant getter and setter.

https://jira.prplfoundation.org/browse/PPM-2182

Signed-off-by: Benjamin Carlson <b.carlson@cablelabs.com>