build: Add ENABLE_VBSS build flag for optional VBSS support
Currently there is no way to enable or disable the implementation of
Virtual BSS.

This flag is necessary because Virtual BSS is an optional addition to
the EasyMesh spec and therefore must be able to be opt-in to the
implementation.

This flag is reflected also in the CI files to enable VBSS code
compilation in the CI pipeline.

https://jira.prplfoundation.org/browse/PPM-2182

Signed-off-by: Benjamin Carlson <b.carlson@cablelabs.com>