common: hostapd: configuration: add config_id parameter
In order to reconfigure only the BSSes that have changed, compute and
write config_id to the configuration file.

The config_id we compute is a sha256 digest of all the BSS parameters.

Note that in prplOS (and OpenWrt) a md5sum is used instead, but it
doesn't matter: the config_id just has to change when one of the BSS
parameters changes.

With this change, hostapd will only reconfigure the BSSes that had
their config_id changed.

Note that currently we don't do anything to try to match config ids,
and as a consequence:

- The order and numbering of the BSSes matter. For example with 3
BSSes, removing the second one will trigger the third one to be
reconfigured (as the current logic fills in the hostapd BSSes
sequentially, the third one will become the second one in the list, so
in practice the config_id of the second BSS will change and clients
will be disconnected).

- When renewing the configuration, the prplMesh controller sends the
BSSes in the same order as it originally did. There is however no
guarantee about it in the standard, which means non-prplMesh
controllers could change the order from one configuration to the next
one, leading to a disconnection of clients.

One way to address those short-comings would be to try to match
specific parts of the existing BSSes configurations (e.g. the SSID,
etc) to re-use the ones that have the same settings. This is however
left for the future (see PPM-2248).

Also note that this feature is currently not implemented for
dwpal-based platforms. For that, PPM-2236 has been created.

Closes PPM-1793
Closes PPM-1889
Closes PPM-2188

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>