bwl: Remove attachment to VAP interfaces on Station mode
The station does not use VAPS.
On the base class of DWPAL and DWPALD we have a state "AttachVaps"
which we don't have on other BWL flavors.
We jump to this state always after the state of "GetRadioInfo".
On station mode, we don't need to attach VAPS, therefore, we shouldn't
do it on that mode.
Added a code that prevents attaching to the VAPs in station mode, and
switch right away to the next state.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>