bwl: Fix context override on station mode of dwpald
In the AP implementation of the DWPALD, we save the context of the AP
object and pass it to the DWAPLD client when attaching and registering
to events.
The code of the AP was copied to the station implementation, but we
can't use it since it causing a bug.
Since all station objects run on the same process, different from the AP
implementation which each AP object runs on a different process, what
happens is that when a newer station object is created, it overrides the
previous station object context, thus when there is an event for the
previous station object, it will be received in the newer station
object.

To fix it, change the context we save to a map whose key is the
the interface name.
When the event will be received, it will search for the right context to
send the message to.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>